requirements.txt
streamlit
langchain
langchain_community
langchain_groq
python-dotenv
duckduckgo-search
validators
nltk
